moderator feature

+++ app/routers/user.py
@@ -18,14 +18,14 @@ router = APIRouter(
 
 
 @router.get('/')
-def fetch_users(skip: int = 0, limit: int = 100, db: Session = Depends(get_db), current_user: int = Depends(get_current_user)):
+def fetch_users(skip: int = 0, limit: int = 100, db: Session = Depends(get_db), current_user = Depends(get_current_user)):
     """ List to get all users """
 
     return crud.get_users(db, skip=skip, limit=limit)
 
 
 @router.get('/{username}')
-def fetch_user(username: str, db: Session = Depends(get_db), current_user: int = Depends(get_current_user)):
+def fetch_user(username: str, db: Session = Depends(get_db), current_user = Depends(get_current_user)):
     """ Fetch a user by username  """
 
     user = crud.get_user(db, username=username)
@@ -36,7 +36,7 @@ def fetch_user(username: str, db: Session = Depends(get_db), current_user: int =
 
 
 @router.patch('/edit/{username}')
-def update_user(user: schema.UserUpdate, username: str, db: Session = Depends(get_db), current_user: int = Depends(get_current_user)):
+def update_user(user: schema.UserUpdate, username: str, db: Session = Depends(get_db), current_user = Depends(get_current_user)):
     """ Update a User profile by username  """
 
     user_db = db.query(User).filter(User.username == username).first()
@@ -64,18 +64,20 @@ def update_user(user: schema.UserUpdate, username: str, db: Session = Depends(ge
 
 
 @router.delete('/delete/{username}/{user_id}')
-def delete_user(username: str, user_id: int, db: Session = Depends(get_db)):
+def delete_user(username: str, user_id: int, db: Session = Depends(get_db),current_user = Depends(get_current_user) ):
     """ Delete a user by it's username  """
-    try:
-        delete_user = crud.delete_user(
-            db, username=username, current_user=user_id)
-        if not delete_user:
-            raise HTTPException(
-                status_code=404, detail=f"user with user_id : {username} does not exist")
-        return {"success": True, "data": "User has been deleted successfully"}
-    except:
+   
+    if user_id == current_user.user_id or current_user.is_admin:
+        try:
+            delete_user = crud.delete_user(
+                db, username=username, current_user=user_id)
+            if not delete_user:
+                raise HTTPException(
+                    status_code=404, detail=f"user with user_id : {username} does not exist")
+            return {"success": True, "data": "User has been deleted successfully"}
+        except:
 
-        return {"error" : "Unable to delete user"}
+            return {"error" : "Unable to delete user"}
